backend/requirements.txt
# WealthPath AI Backend Dependencies

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client & External APIs
httpx==0.25.2
aiohttp==3.9.1

# ML & Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2

# LLM Providers
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.8.0

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2
tiktoken==0.5.2

# Utilities
click==8.1.7
python-slugify==8.0.1
celery==5.3.4